Add conditional post name field
The code now inserts the user's inputted name for a post if the user is authenticated. The 'name' field in the 'StorePostRequest' has been made nullable and validated with the 'AllowedIfAuthenticated' rule. This change benefits the CreatePost action by inserting the potentially nullable 'name' into the post if the user is authenticated.